+++ Hex/Resources/changelog.md
@@ -1,3 +1,5 @@
+- Rework audio pausing logic to make it more reliable
+
 ### 0.1.26
 - Add changelog
 - Add option to set minimum record time